Feature :: Favorite Scan

+++ app/src/main/java/com/example/tastetrove/helper/ImageClassifierHelper.kt
@@ -61,7 +61,7 @@ class ImageClassifierHelper(private val listener: ImageClassifierListener?) {
 
         val imageProcessor = ImageProcessor.Builder()
             .add(ResizeOp(224, 224, ResizeOp.ResizeMethod.NEAREST_NEIGHBOR))
-            .add(CastOp(DataType.UINT8))
+            .add(CastOp(DataType.FLOAT32))
             .build()
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
